requirements.txt
Flask==2.1.2
pydub==0.25.1
pyannote.audio==2.1.1
torch==1.12.0
numpy==1.23.1
